refactor: module structure, starter commands refactored & documentation refactored

+++ link/uninstall.sh
@@ -5,8 +5,10 @@ source $DOTFILES_PATH/lib/utils.sh
 start_msg "removing symlinks..."
 
 for i in $(list_link_files); do
-    info_msg "removing $HOME/$i"
-    rm -rf $HOME/$i
+    if [ -f "$HOME/$i" ]; then
+        info_msg "removing $HOME/$i"
+        rm -rf $HOME/$i
+    fi
 done;
 
 done_msg "removing symlinks"
